Добавлены изменения в сохранение offer`а и сделано получение всех кат…

+++ src/main/java/com/gb/agile/craft_master/repositories/ServiceRepository.java
@@ -3,5 +3,5 @@ package com.gb.agile.craft_master.repositories;
 import com.gb.agile.craft_master.model.entities.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Service, Long> {
+public interface ServiceRepository extends JpaRepository<Service, Integer> {
 }
